CF symbology renderer
