+++ VSProjects/Geomapmaker/Data/CFSymbology.cs
@@ -76,67 +76,5 @@ namespace Geomapmaker.Data
             CFSymbolOptionsList = cfSymbols;
         }
 
-        public static async Task AddSymbolToRenderer(string key, string symbolJson)
-        {
-            // Find the ContactsFaults layer
-            FeatureLayer layer = MapView.Active.Map.GetLayersAsFlattenedList().OfType<FeatureLayer>().FirstOrDefault(l => l.Name == "ContactsAndFaults");
-
-            if (layer == null)
-            {
-                return;
-            }
-
-            await QueuedTask.Run(() =>
-            {
-                //
-                // Update Renderer
-                //
-
-                CIMUniqueValueRenderer layerRenderer = layer.GetRenderer() as CIMUniqueValueRenderer;
-
-                CIMUniqueValueGroup layerGroup = layerRenderer?.Groups?.FirstOrDefault();
-
-                List<CIMUniqueValueClass> listUniqueValueClasses = layerGroup == null ? new List<CIMUniqueValueClass>() : new List<CIMUniqueValueClass>(layerGroup.Classes);
-
-                // Check if the renderer already has symbology for that key
-                if (listUniqueValueClasses.Any(a => a.Label == key))
-                {
-                    return;
-                }
-
-                CIMUniqueValue[] listUniqueValues = new CIMUniqueValue[] {
-                        new CIMUniqueValue {
-                            FieldValues = new string[] { key }
-                        }
-                    };
-
-                CIMUniqueValueClass uniqueValueClass = new CIMUniqueValueClass
-                {
-                    Editable = false,
-                    Label = key,
-                    Description = key,
-                    Patch = PatchShape.AreaPolygon,
-                    Symbol = CIMSymbolReference.FromJson(symbolJson, null),
-                    Visible = true,
-                    Values = listUniqueValues,
-                };
-                listUniqueValueClasses.Add(uniqueValueClass);
-
-                CIMUniqueValueGroup uvg = new CIMUniqueValueGroup
-                {
-                    Classes = listUniqueValueClasses.ToArray(),
-                };
-                CIMUniqueValueGroup[] listUniqueValueGroups = new CIMUniqueValueGroup[] { uvg };
-
-                CIMUniqueValueRenderer updatedRenderer = new CIMUniqueValueRenderer
-                {
-                    UseDefaultSymbol = false,
-                    Groups = listUniqueValueGroups,
-                    Fields = new string[] { "symbol" }
-                };
-
-                layer.SetRenderer(updatedRenderer);
-            });
-        }
     }
 }
